build-and-test.sh: Add build-and-test.sh
build-and-test.sh does what was previously done only in the .travis.yml.
It does sed substitution on the Dockerfile to build based on a
particular distro, as well as runs the test.

That way a user can test things locally just like travis does, since
travis now will invoke build-and-test.sh.

Signed-off-by: Randy Witt <randy.e.witt@linux.intel.com>