Fix runtests.sh failure when SELinux is enabled
When using Docker with SELinux enabled on the host, volume mounted files
and directories must be properly labelled. Correct labelling can be
performed automatically by appending a :Z to the docker run volume mount
option. This change updates the workdir volume mount option in
runtests.sh to automatically label it properly for SELinux.

Signed-off-by: John Akre <john.w.akre@intel.com>