Add run-workdir-check.sh
Since the workdir can now be specified using an argument to the
entrypoint or as an argument to docker, we need to test that the workdir
is as expected in the container for either case.

run-workdir-check.sh allows us to do the check for both when runtests.sh
sees that run-workdir-check.sh is the test being executed.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>

squash

Signed-off-by: Randy Witt <randy.e.witt@intel.com>