sudoers.usersetup: Allow usersetup to preserve PATH
If the PATH was set using "ENV" in the Dockerfile, it would not be
preserved. This is due to the fact that "secure_path" is set in the
sudoers file.

Since a user may want to build a Docker image with a modified path

pokyuser, allow usersetup to preserve the path when it runs sudo.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>