Fix update to Ubuntu 18.04 as the default
The user is now required to specify the BASE_DISTRO as a build-arg when
using the Dockerfile. .travis.yml has also been updated to use
Ubuntu 18.04 as the default.

The previous commit in 7389f739a825ed14db499194f693eb5b6deb17fb to
change the default to Ubuntu 18.04 missed some additional changes needed
to actually change to Ubuntu 18.04. Because of this all images contained
Ubuntu 18.04 instead of the desired distro.

Requiring the user to specify the BASE_DISTRO will hopefully prevent
this kind of mistake in the future.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>