Dockerfile: Create the /home/yoctouser directory
The user yoctouser is removed and the directory will be removed by the
userdel command as well. However, when using docker to build the
containers, the /home/yoctouser would get created again implicitly due
to WORKDIR being set to /home/yoctouser. The existence of this directory
is used in some tests.

Some container runtimes such as the one used by podman will not create
the WORKDIR directory. So for now manually create an empty
/home/yoctouser since it is used for some checks.

This makes builds with podman work by mimicking the docker behavior.
Ideally the dependency on this behavior should be removed.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>