distro-check.sh: add debian-12 and fedora-40
This is easy to forget to add.

Thank you again to Rob Woolley for adding these two new
distributions to the matrix.

FIXME:
Since the release date is April 12, 2024, the fedora-40
container is reporting "Prerelease" so we need to satisify
that string for now:

PRETTY_NAME="Fedora Linux 40 (Container Image Prerelease)"

Signed-off-by: Tim Orling <tim.orling@konsulko.com>